requirements.txt
django
gunicorn
django-heroku
whitenoise
numpy
opencv-python-headless
pillow
imutils
PyPDF2